+++ miracle/api/tests.py
@@ -1,3 +1,5 @@
+import json
+
 from miracle.util import gzip_encode
 
 
@@ -29,6 +31,12 @@ def test_delete_fail(app, stats):
     ])
 
 
+def test_jwk(app, stats):
+    res = app.get('/v1/jwk', status=200)
+    assert set(res.json.keys()) == {'n', 'e', 'kty'}
+    assert res.json['kty'] == 'RSA'
+
+
 def test_stats(app, stats):
     res = app.get('/v1/stats', status=200)
     assert CORS_HEADERS - set(res.headers.keys()) == set()
@@ -38,34 +46,44 @@ def test_stats(app, stats):
     ])
 
 
-def test_upload(app, stats):
-    res = app.post_json('/v1/upload',
-                        {'foo': 1},
-                        headers={'Content-Type': 'application/json',
-                                 'X-User': b'abc'},
-                        status=200)
+def test_upload(app, crypto, stats):
+    data = crypto.encrypt(json.dumps({'foo': 1}))
+    res = app.post('/v1/upload',
+                   data,
+                   headers={'Content-Type': 'text/plain',
+                            'X-User': b'abc'},
+                   status=200)
     assert CORS_HEADERS - set(res.headers.keys()) == set()
     assert res.json == {'status': 'success'}
-    app.post('/v1/upload', b'no json',
-             headers={'Content-Type': 'application/json',
+
+    data = crypto.encrypt(b'encrypted no json')
+    app.post('/v1/upload', data,
+             headers={'Content-Type': 'text/plain',
                       'X-User': b'abc'},
              status=200)
+
+    data = b'no json'
+    app.post('/v1/upload', data,
+             headers={'Content-Type': 'text/plain',
+                      'X-User': b'abc'},
+             status=200)
+
     stats.check(timer=[
-        ('task', 2, ['task:data.tasks.upload']),
+        ('task', 3, ['task:data.tasks.upload']),
     ])
 
 
 def test_upload_fail(app, stats):
     app.post('/v1/upload', b'',
-             headers={'Content-Type': 'application/json'},
+             headers={'Content-Type': 'text/plain'},
              status=400)
     app.post('/v1/upload', b'',
-             headers={'Content-Type': 'application/json',
+             headers={'Content-Type': 'text/plain',
                       'X-User': b'abc'},
              status=400)
     too_large = b'a' + b'0123456789' * 1024 * 1024
     app.post('/v1/upload', too_large,
-             headers={'Content-Type': 'application/json',
+             headers={'Content-Type': 'text/plain',
                       'X-User': b'abc'},
              status=400)
     stats.check(timer=[
@@ -73,11 +91,12 @@ def test_upload_fail(app, stats):
     ])
 
 
-def test_upload_gzip(app):
+def test_upload_gzip(app, crypto):
+    data = gzip_encode(crypto.encrypt(b'{"foo": 1}'))
     res = app.post('/v1/upload',
-                   gzip_encode('{"foo": 1}'),
+                   data,
                    headers={'Content-Encoding': 'gzip',
-                            'Content-Type': 'application/json',
+                            'Content-Type': 'text/plain',
                             'X-User': b'abc'},
                    status=200)
     assert res.json == {'status': 'success'}
@@ -87,7 +106,7 @@ def test_upload_gzip_fail(app):
     app.post('/v1/upload',
              b'invalid',
              headers={'Content-Encoding': 'gzip',
-                      'Content-Type': 'application/json',
+                      'Content-Type': 'text/plain',
                       'X-User': b'abc'},
              status=400)